Permission change + Improve &(function)me

+++ cmd/sayd.js
@@ -1,7 +1,10 @@
 module.exports.run = (client, message, args) => {
-	let sayMessage = args.join(" ");
-	message.author.lastMessage.delete();
-	message.channel.send(sayMessage);
+	if (message.member.roles.find("name", "Owner")) {
+		let sayMessage = args.join(" ");
+		message.author.lastMessage.delete();
+		message.channel.send(sayMessage);
+	}
+	else message.channel.send("You don't have enough permission to use this :3");
 }
 
 module.exports.help = {
